DeepSeek/back_DeepSeek/requirements.txt
fastapi
uvicorn
pandas
openpyxl
pyodbc
python-dotenv
requests
yfinance
uuid